GetRankingResponseの変更に伴い、ランキング関連の型を修正
