+++ src/app/_types/ranking.ts
@@ -6,3 +6,21 @@ export interface Ranking {
   rank: number;
   value: number;
 }
+
+export interface RankingData {
+  dailyRankingByCount: Ranking[];
+  dailyRankingByTime: Ranking[];
+  monthlyRankingByCount: Ranking[];
+  monthlyRankingByTime: Ranking[];
+  weeklyRankingByCount: Ranking[];
+  weeklyRankingByTime: Ranking[];
+  yearlyRankingByCount: Ranking[];
+  yearlyRankingByTime: Ranking[];
+}
+
+/** ランキング取得APIのレスポンス型定義 */
+export interface GetRankingResponse {
+  status: "success";
+  message: string;
+  data: RankingData;
+}
