--- src/app/api/ranking/_types/response.ts
@@ -1,17 +1,0 @@
-import { Ranking } from "@/app/_types/ranking";
-
-/** ランキングデータのレスポンス型定義 */
-export interface RankingResponseType {
-  status: string;
-  message: string;
-  data: {
-    dailyRankingByCount: Ranking[];
-    dailyRankingByTime: Ranking[];
-    monthlyRankingByCount: Ranking[];
-    monthlyRankingByTime: Ranking[];
-    weeklyRankingByCount: Ranking[];
-    weeklyRankingByTime: Ranking[];
-    yearlyRankingByCount: Ranking[];
-    yearlyRankingByTime: Ranking[];
-  };
-}
